Adjusted pages to include components and containers; file cleanup finished. Moving to styling.
