+++ src/pages/account/account.tsx
@@ -1,4 +1,4 @@
-import AccountContainer from "containers/account/AccountContainer";
+import AccountContainer from "@/src/pages/account/containers/AccountContainer";
 
 
 export default function AccountPage() {
